#166 Changes due to moving observation_base.py
