+++ necst/procedures/observations/otf.py
@@ -4,7 +4,7 @@ from typing import Optional, Tuple, Union
 
 from neclib.parameters import ObsParams
 
-from .observation_base import Observation
+from ..observation_base import Observation
 
 
 class OTF(Observation):
