+++ necst/procedures/observations/skydip.py
@@ -3,7 +3,7 @@ from typing import Union
 from neclib.parameters import PointingError
 
 from ... import config
-from .observation_base import Observation
+from ..observation_base import Observation
 
 
 class Skydip(Observation):
